build(sdl3): migrate entire codebase from SDL2 to SDL3
- Update build system (nob.c) to use pkg-config for SDL3 libraries
- Replace all SDL2 headers with SDL3 equivalents across src/, examples/, tests/
- Convert SDL2 API calls to SDL3: event types, rendering functions, texture/surface APIs
- Replace SDL_gfx rounded rectangle drawing with SDL_RenderGeometry-based implementations
- Adapt SDLApp and Window constructors to match SDL3 signature changes
- Update test helpers to use new SDL3 event structure
- Remove SDL2-specific initialization (IMG_Init) and dependencies
- Add SDL3 migration plan document
- Delete obsolete files (TODO.md, memory-bank-instructions.md, unused asset)

+++ src/timer_manager.cpp
@@ -23,7 +23,7 @@ void TimerManager::removeTimer(uint32_t timerId) {
 }
 
 void TimerManager::update() {
-    uint32_t currentTime = SDL_GetTicks();
+    Uint64 currentTime = SDL_GetTicks();
 
     // Collect timers that need to execute, to avoid iterator invalidation
     // if callbacks modify the timers vector
